requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.1
django-unfold==0.21.1
gunicorn==21.2.0
idna==3.6
packaging==24.0
pafy==0.5.5
Pillow==9.5.0
pytube==15.0.0
pytz==2024.1
requests==2.31.0
soupsieve==2.5
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.2.1
waitress==3.0.0
whitenoise==6.6.0
